Make sure the whites and yellows are using the default terminal color
This is an issue I get on iTerm2 where I have a light color theme, which
is preventing me from being able to see any color because it's using the
same color values. With this commit, it will use default colors which
will honor the color values of the presets in iTerm.